Add TripResponseMapper
